% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{brauer_gene_exp}
\alias{brauer_gene_exp}
\title{Gene expression in yeast measured by microarray}
\format{
A data frame with 198,430 observations of 7 variables.
}
\source{
\url{http://www.molbiolcell.org/content/19/1/352.long}

\url{http://varianceexplained.org/r/tidy-genomics/}

\url{http://varianceexplained.org/r/tidy-genomics-broom/}
}
\usage{
brauer_gene_exp
}
\description{
A dataset containing gene expression values of Saccharomyces cerevisiae grown
under nutrient limitation conditions. The variables are as follows:
}
\details{
\itemize{
\item \code{name} Common gene name
\item \code{BP} Gene Ontology Biological Process
\item \code{MF} Gene Ontology Molecular Function
\item \code{systematic_name} Systematic gene name
\item \code{nutrient} Nutrient omitted for growth
\item \code{rate} Growth rate
\item \code{expression} Gene expression value
}
}
\keyword{datasets}
