% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clust_mat}
\alias{clust_mat}
\title{Example matrix to highlight clustering methods}
\format{matrix with 4 rows and 8 columns}
\usage{
clust_mat
}
\description{
Example matrix to highlight clustering methods
}
\examples{
clust_mat

}
\keyword{datasets}
