% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.r
\docType{package}
\name{eda}
\alias{eda}
\alias{eda-package}
\title{Exploratory Data Analysis in R/RStudio}
\description{
Provides course materials for IDPT 7620, Practical Biological Data Analysis in R.
}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@gmail.com} (0000-0002-6299-179X)

Authors:
\itemize{
  \item Austin Gillen \email{austin.gillen@gmail.com}
  \item Kent Riemondy \email{kent.riemondy@gmail.com}
}

}
