% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{esc_mat}
\alias{esc_mat}
\title{single cell RNA-seq dataset of mouse embryonic development}
\format{matrix with 22431 genes as rows and 268 cells as columns containing
raw read counts.}
\usage{
esc_mat
}
\description{
A scRNA-seq dataset of developing mouse embryos generated with Smart-Seq2
}
\examples{
esc_mat

}
\keyword{datasets}
