% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{missing_ex}
\alias{missing_ex}
\title{Sample missing data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 5 columns.
}
\usage{
missing_ex
}
\description{
A dataset for exploring missing data.
}
\keyword{datasets}
