% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mtcars_tbl}
\alias{mtcars_tbl}
\title{mtcars in tibble format}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 12 columns.
}
\usage{
mtcars_tbl
}
\description{
mtcars in tibble format
}
\examples{
mtcars_tbl

}
\keyword{datasets}
