% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nci60}
\alias{nci60}
\title{Microarray data from the NCI-60 cancer cell line panel}
\format{matrix with 6830 rows and 64 columns containing
normalized gene expression values.}
\source{
ISLR::NCI60
}
\usage{
nci60
}
\description{
See ?ISLR::NCI60 for description of dataset.
}
\examples{
nci60

}
\keyword{datasets}
