% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbda.r
\docType{package}
\name{pbda}
\alias{pbda}
\alias{pbda-package}
\title{Practical Biological Data Analysis in R/RStudio}
\description{
Provides course materials for IDPT 7810, Practical Biological Data Analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IDPT7810/practical-data-analysis}
  \item \url{https://idpt7810.github.io/practical-data-analysis/}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@gmail.com} (0000-0002-6299-179X)

Authors:
\itemize{
  \item Austin Gillen \email{austin.gillen@ucdenver.edu}
  \item Kent Riemondy \email{kent.riemondy@ucdenver.edu}
  \item Rui Fu \email{rui.fu@ucdenver.edu}
}

}
