% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{qpcr}
\alias{qpcr}
\title{Example Quantitative PCR data}
\format{A tibble containing 145 observations of 6 variables.
\describe{
\item{well}{plate well}
\item{fluor}{Fluorophore}
\item{target}{mRNA target}
\item{content}{Not sure}
\item{sample}{sample name}
\item{cq}{Quantification cycle}
}}
\source{
Rachel Ancar
}
\usage{
qpcr
}
\description{
A dataset containing Bio-Rad qPCR data for IFNbeta and Actin upon
transfection with RNA ligands.
}
\keyword{datasets}
