% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{qpcr_data}
\alias{qpcr_data}
\title{Names of example qPCR data}
\format{
qPCR sample data in 96 well format.
}
\usage{
qpcr_data
}
\description{
Relative expression levels for each sample.
}
\examples{
qpcr_data

}
\keyword{datasets}
