% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tx_rates}
\alias{tx_rates}
\title{Transcription rates for ~7000 genes measured across a timecourse}
\format{
matrix with 7872 genes as rows and 13 columns containing
transcription rates.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE56977}

\url{https://www.ncbi.nlm.nih.gov/pubmed/25497548}
}
\usage{
tx_rates
}
\description{
Mouse dendritic cells were stimulated with LPS, and timepoints were
collected every 15 minutes. Metabolic labeling pulse with 4sU was performed
10 minutes prior to collection. Total RNA and 4sU containing RNA was collected
and used to calculate transcription rates.
}
\examples{
tx_rates

}
\keyword{datasets}
