% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-eda.R
\name{update_eda}
\alias{update_eda}
\title{Update eda package using devtools}
\usage{
update_eda(repo = "IDPT7810/practical-data-analysis", ...)
}
\arguments{
\item{repo}{repository to update}

\item{...}{additional arguments to pass to devtools::install_github()}
}
\description{
Update eda package using devtools
}
\examples{
\dontrun{
update_eda()
}
}
