% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-pbda.R
\name{update_pbda}
\alias{update_pbda}
\title{Update pbda package using devtools}
\usage{
update_pbda(repo = "IDPT7810/practical-data-analysis", ...)
}
\arguments{
\item{repo}{repository to update}

\item{...}{additional arguments to pass to devtools::install_github()}
}
\description{
Update pbda package using devtools
}
\examples{
\dontrun{
update_pbda()
}
}
