% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{yeast_prot_prop}
\alias{yeast_prot_prop}
\title{Properties of yeast proteins}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6704 rows and 40 columns.}
\source{
\url{https://downloads.yeastgenome.org/curation/calculated_protein_info/protein_properties.tab}
}
\usage{
yeast_prot_prop
}
\description{
A dataset containing physical parameters of yeast proteins.``
}
\keyword{datasets}
