% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbda.r
\docType{package}
\name{pbda}
\alias{pbda}
\alias{pbda-package}
\title{Practical Biological Data Analysis in R/RStudio}
\description{
Provides course materials for IDPT 7810, Practical Biological Data Analysis in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{h https://rnabioco.github.io/practical-data-analysis}
  \item \url{https://github.com/rnabioco/practical-data-analysis}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@gmail.com} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})

Authors:
\itemize{
  \item Austin Gillen \email{austin.gillen@cuanschutz.edu}
  \item Kent Riemondy \email{kent.riemondy@cuanschutz.edu}
  \item Rui Fu \email{rui.fu@cuanschutz.edu}
  \item Ryan Sheridan \email{ryan.sheridan@cuanschutz.edu}
}

}
