% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{qpcr_data}
\alias{qpcr_data}
\title{Example qPCR data}
\format{qPCR sample data in 96 well format.}
\usage{
qpcr_data
}
\description{
Example qPCR data
}
\examples{
qpcr_data

}
\keyword{datasets}
